requirements.txt
python-dotenv
aiogram==3.22.0
sqlalchemy==2.0.23
aiosqlite==0.19.0